Added the dgeq app.
It scrapes DGEQ's site for contributions. It could, instead, just import the
CSV export it provides, but if it did, we wouldn't have the "idrech" info which
is a unique identifier for each name which we don't have in the CSV export.

--HG--
rename : assnat/management/commands/scrape.py => assnat/management/commands/scrapeassnat.py