Improved OIQ scraping.
Instead of trying to link with base models (Person, Company) early, it simply
scrapes as much information as it can and records raw data.

Also, when searching for a name, instead of trying to find the right record
to scrape (in case of multiple results), it simply scrapes all results in the
table.